+++ src/test/java/com/excercise/tripaths/GraphTestWrapper.java
@@ -31,7 +31,7 @@ public class GraphTestWrapper {
     /**
      * @apiNote use this method only when the given vertex has neighbours with unique weight.
      */
-    public Map<Integer, WeightedVertex> getNeighbourListOfByWeight(final WeightedVertex vertex) {
+    public Map<Long, WeightedVertex> getNeighbourListOfByWeight(final WeightedVertex vertex) {
         return Graphs.neighborListOf(wrapped, vertex).stream().collect(Collectors.toMap(WeightedVertex::getWeight, Function.identity()));
     }
